move country names to new 'de' base
