+++ faker/providers/address/de/__init__.py
@@ -1,0 +1,66 @@
+# coding=utf-8
+
+from __future__ import unicode_literals
+from .. import Provider as AddressProvider
+
+
+class Provider(AddressProvider):
+    countries = (
+        'Afghanistan', 'Alandinseln', 'Albanien', 'Algerien',
+        'Amerikanisch-Ozeanien', 'Amerikanisch-Samoa',
+        'Amerikanische Jungferninseln', 'Andorra', 'Angola', 'Anguilla',
+        'Antarktis', 'Antigua und Barbuda', 'Argentinien', 'Armenien', 'Aruba',
+        'Aserbaidschan', 'Australien', 'Bahamas', 'Bahrain', 'Bangladesch',
+        'Barbados', 'Belarus', 'Belgien', 'Belize', 'Benin', 'Bermuda',
+        'Bhutan', 'Bolivien', 'Bosnien und Herzegowina', 'Botsuana',
+        'Bouvetinsel', 'Brasilien', 'Britische Jungferninseln',
+        'Britisches Territorium im Indischen Ozean', 'Brunei Darussalam',
+        'Bulgarien', 'Burkina Faso', 'Burundi', 'Chile', 'China', 'Cookinseln',
+        'Costa Rica', 'Côte d’Ivoire', 'Demokratische Republik Kongo',
+        'Demokratische Volksrepublik Korea', 'Deutschland', 'Dominica',
+        'Dominikanische Republik', 'Dschibuti', 'Dänemark', 'Ecuador',
+        'El Salvador', 'Eritrea', 'Estland', 'Falklandinseln', 'Fidschi',
+        'Finnland', 'Frankreich', 'Französisch-Guayana',
+        'Französisch-Polynesien', 'Färöer', 'Gabun', 'Gambia', 'Georgien',
+        'Ghana', 'Gibraltar', 'Grenada', 'Griechenland', 'Grönland',
+        'Guadeloupe', 'Guam', 'Guatemala', 'Guernsey', 'Guinea',
+        'Guinea-Bissau', 'Guyana', 'Haiti', 'Heard- und McDonald-Inseln',
+        'Honduras', 'Indien', 'Indonesien', 'Irak', 'Iran', 'Irland', 'Island',
+        'Isle of Man', 'Israel', 'Italien', 'Jamaika', 'Japan', 'Jemen',
+        'Jersey', 'Jordanien', 'Kaimaninseln', 'Kambodscha', 'Kamerun',
+        'Kanada', 'Kap Verde', 'Kasachstan', 'Katar', 'Kenia', 'Kirgisistan',
+        'Kiribati', 'Kokosinseln', 'Kolumbien', 'Komoren', 'Kongo', 'Kroatien',
+        'Kuba', 'Kuwait', 'Laos', 'Lesotho', 'Lettland', 'Libanon', 'Liberia',
+        'Libyen', 'Liechtenstein', 'Litauen', 'Luxemburg', 'Madagaskar',
+        'Malawi', 'Malaysia', 'Malediven', 'Mali', 'Malta', 'Marokko',
+        'Marshallinseln', 'Martinique', 'Mauretanien', 'Mauritius', 'Mayotte',
+        'Mazedonien', 'Mexiko', 'Mikronesien', 'Monaco', 'Mongolei',
+        'Montenegro', 'Montserrat', 'Mosambik', 'Myanmar', 'Namibia', 'Nauru',
+        'Nepal', 'Neukaledonien', 'Neuseeland', 'Nicaragua', 'Niederlande',
+        'Niederländische Antillen', 'Niger', 'Nigeria', 'Niue', 'Norfolkinsel',
+        'Norwegen', 'Nördliche Marianen', 'Oman', 'Osttimor', 'Pakistan',
+        'Palau', 'Palästinensische Gebiete', 'Panama', 'Papua-Neuguinea',
+        'Paraguay', 'Peru', 'Philippinen', 'Pitcairn', 'Polen', 'Portugal',
+        'Puerto Rico', 'Republik Korea', 'Republik Moldau', 'Ruanda',
+        'Rumänien', 'Russische Föderation', 'Réunion', 'Salomonen', 'Sambia',
+        'Samoa', 'San Marino', 'Saudi-Arabien', 'Schweden', 'Schweiz',
+        'Senegal', 'Serbien', 'Serbien und Montenegro', 'Seychellen',
+        'Sierra Leone', 'Simbabwe', 'Singapur', 'Slowakei', 'Slowenien',
+        'Somalia', 'Sonderverwaltungszone Hongkong',
+        'Sonderverwaltungszone Macao', 'Spanien', 'Sri Lanka',
+        'St. Barthélemy', 'St. Helena', 'St. Kitts und Nevis', 'St. Lucia',
+        'St. Martin', 'St. Pierre und Miquelon',
+        'St. Vincent und die Grenadinen', 'Sudan', 'Suriname',
+        'Svalbard und Jan Mayen', 'Swasiland', 'Syrien',
+        'São Tomé und Príncipe', 'Südafrika',
+        'Südgeorgien und die Südlichen Sandwichinseln', 'Tadschikistan',
+        'Taiwan', 'Tansania', 'Thailand', 'Togo', 'Tokelau', 'Tonga',
+        'Trinidad und Tobago', 'Tschad', 'Tschechische Republik', 'Tunesien',
+        'Turkmenistan', 'Turks- und Caicosinseln', 'Tuvalu', 'Türkei',
+        'Uganda', 'Ukraine', 'Ungarn', 'Uruguay', 'Usbekistan', 'Vanuatu',
+        'Vatikanstadt', 'Venezuela', 'Vereinigte Arabische Emirate',
+        'Vereinigte Staaten', 'Vereinigtes Königreich', 'Vietnam',
+        'Wallis und Futuna', 'Weihnachtsinsel', 'Westsahara',
+        'Zentralafrikanische Republik', 'Zypern', 'Ägypten',
+        'Äquatorialguinea', 'Äthiopien', 'Äußeres Ozeanien', 'Österreich',
+    )
